+++ src/containers/Root.js
@@ -1,14 +1,36 @@
 import React, { Component } from 'react';
+import PropTypes from 'prop-types';
+import { withRouter } from 'react-router';
 import ReactGA from 'react-ga';
 import ReactPixel from 'react-facebook-pixel';
 import { ScrollContext } from 'react-router-scroll-4';
 import App from '../components/Layout';
 
+const logPageView = location => {
+  ReactGA.set({ page: location.pathname + location.search });
+  ReactGA.pageview(location.pathname + location.search);
+};
 
 class Root extends Component {
   constructor(props) {
     super(props);
     this.init();
+  }
+
+  componentDidMount() {
+    const {
+      location,
+    } = this.props;
+
+    logPageView(location);
+  }
+
+  componentDidUpdate() {
+    const {
+      location,
+    } = this.props;
+
+    logPageView(location);
   }
 
   init = () => {
@@ -29,4 +51,12 @@ class Root extends Component {
   }
 }
 
-export default Root;
+Root.propTypes = {
+  location: PropTypes.shape({
+    pathname: PropTypes.string,
+    search: PropTypes.string,
+  }),
+};
+
+
+export default withRouter(Root);
